Refactor folder system
